projetos/RPAPetz/requirements.txt

botcity-maestro-sdk>=0.3.3,<1.0
pdfplumber
pandas
openai
# Uncomment to use botcity-utils
# botcity-utils>=0.1.2

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
# pandas
